+++ tests/ats/test_e2e.py
@@ -11,7 +11,7 @@ import yaml
 from pykube.objects import NamespacedAPIObject
 from pytest_helm_charts.clusters import Cluster, ExistingCluster
 from pytest_helm_charts.flux.git_repository import make_git_repository_obj, wait_for_git_repositories_to_be_ready
-from pytest_helm_charts.giantswarm_app_platform.app import make_app_object, ConfiguredApp
+from pytest_helm_charts.giantswarm_app_platform.app import make_app_object
 from pytest_helm_charts.giantswarm_app_platform.catalog import make_catalog_obj
 from pytest_helm_charts.k8s.deployment import wait_for_deployments_to_run
 from pytest_helm_charts.utils import wait_for_objects_condition
@@ -19,13 +19,28 @@ from yaml import SafeLoader
 
 logger = logging.getLogger(__name__)
 
+
 class NamespacedKonfigureOperatorCR(NamespacedAPIObject, abc.ABC):
     pass
+
+
+class KonfigurationSchemaCR(NamespacedKonfigureOperatorCR):
+    version = "konfigure.giantswarm.io/v1alpha1"
+    endpoint = "konfigurationschemas"
+    kind = "KonfigurationSchema"
+
+
+class KonfigurationCR(NamespacedKonfigureOperatorCR):
+    version = "konfigure.giantswarm.io/v1alpha1"
+    endpoint = "konfigurations"
+    kind = "Konfiguration"
+
 
 class ManagementClusterConfigurationCR(NamespacedKonfigureOperatorCR):
     version = "konfigure.giantswarm.io/v1alpha1"
     endpoint = "managementclusterconfigurations"
     kind = "ManagementClusterConfiguration"
+
 
 @pytest.fixture(scope="module")
 def setup(kube_cluster: Cluster):
@@ -121,17 +136,9 @@ def setup(kube_cluster: Cluster):
     )
 
     # CRDs
-    dir_path = os.path.dirname(os.path.realpath(__file__))
-
-    text_mcc = Path(f"{dir_path}/../../config/crd/bases/konfigure.giantswarm.io_managementclusterconfigurations.yaml").read_text()
-    dict_mcc = yaml.load(text_mcc, SafeLoader)
-
-    crd_mcc = pykube.CustomResourceDefinition(kube_cluster.kube_client, dict_mcc)
-
-    if crd_mcc.exists():
-        crd_mcc.update()
-    else:
-        crd_mcc.create()
+    install_crd(kube_cluster, "../../config/crd/bases/konfigure.giantswarm.io_konfigurations.yaml")
+    install_crd(kube_cluster, "../../config/crd/bases/konfigure.giantswarm.io_konfigurationschemas.yaml")
+    install_crd(kube_cluster, "../../config/crd/bases/konfigure.giantswarm.io_managementclusterconfigurations.yaml")
 
 
 @pytest.mark.functional
@@ -145,6 +152,7 @@ def test_api_working(kube_cluster: Cluster) -> None:
     """
     assert kube_cluster.kube_client is not None
     assert len(pykube.Node.objects(kube_cluster.kube_client)) >= 1
+
 
 @pytest.mark.functional
 def test_mcc_working(setup, kube_cluster: Cluster) -> None:
@@ -210,7 +218,7 @@ def test_mcc_working(setup, kube_cluster: Cluster) -> None:
         ManagementClusterConfigurationCR,
         ["example-1"],
         "default",
-        mcc_cr_ready,
+        cr_ready,
         120,
         False,
     )
@@ -232,7 +240,216 @@ def test_mcc_working(setup, kube_cluster: Cluster) -> None:
     assert secret_values_app_1_ex_1.get("example") == "example"
 
 
-def mcc_cr_ready(cr: NamespacedKonfigureOperatorCR) -> bool:
+raw_schema = """---
+variables:
+  - name: installation
+    required: true
+  - name: app
+    required: true
+layers:
+  - id: defaults
+    path:
+      directory: default
+      required: true
+    values:
+      path:
+        directory: ""
+      configMap:
+        name: config.yaml
+        required: true
+    templates:
+      path:
+        directory: apps/<< app >>
+        required: true
+      configMap:
+        name: configmap-values.yaml.template
+        required: true
+        values:
+          merge:
+            strategy: ConfigMapsInLayerOrder
+      secret:
+        name: secret-values.yaml.template
+        required: false
+        values:
+          merge:
+            strategy: ConfigMapsAndSecretsInLayerOrder
+  - id: management-cluster
+    path:
+      directory: installations/<< installation >>
+      required: true
+    values:
+      path:
+        directory: ""
+      configMap:
+        name: config.yaml.patch
+        required: true
+      secret:
+        name: secret.yaml
+        required: true
+    templates:
+      path:
+        directory: apps/<< app >>
+        required: false
+      configMap:
+        name: configmap-values.yaml.patch
+        required: false
+        values:
+          merge:
+            strategy: ConfigMapsInLayerOrder
+      secret:
+        name: secret-values.yaml.patch
+        required: false
+        values:
+          merge:
+            strategy: SameTypeFromCurrentLayer
+includes:
+  - id: include
+    function:
+      name: include
+    path:
+      directory: include
+      required: true # unused
+    extension: .yaml.template
+  - id: include-self
+    function:
+      name: includeSelf
+    path:
+      directory: include-self
+      required: false # unused
+    extension: .yaml.template
+"""
+
+@pytest.mark.functional
+def test_konfiguration_working(setup, kube_cluster: Cluster) -> None:
+    # THIS IS A MUST TO BE ABLE TO INTERACT WITH CRDs APPLIED DURING TEST COS THE ORIGINAL CLIENT ALREADY DID DISCOVERY
+    # AND IS BEING REUSED ACROSS TESTS
+    fresh_kube_client = ExistingCluster(kube_config_path=kube_cluster.kube_config_path).create()
+
+    konfiguration_schema_cr: Dict[str, Any] = {
+        "apiVersion": KonfigurationSchemaCR.version,
+        "kind": KonfigurationSchemaCR.kind,
+        "metadata": {
+            "name": "example-schema",
+            "namespace": "default",
+        },
+        "spec": {
+            "raw": {
+                "content": raw_schema,
+            },
+        }
+    }
+
+    example_schema = KonfigurationSchemaCR(fresh_kube_client, konfiguration_schema_cr)
+
+    if example_schema.exists():
+        example_schema.update()
+    else:
+        example_schema.create()
+
+    konfiguration_cr: Dict[str, Any] = {
+        "apiVersion": KonfigurationCR.version,
+        "kind": KonfigurationCR.kind,
+        "metadata": {
+            "name": "example-1",
+            "namespace": "default",
+        },
+        "spec": {
+            "targets": {
+                "schema": {
+                    "reference": {
+                        "name": "example-schema",
+                        "namespace": "default"
+                    }
+                },
+                "defaults": {
+                    "variables": [
+                        {
+                            "name": "installation",
+                            "value": "installation-1"
+                        }
+                    ]
+                },
+                "iterations": {
+                    "app-1": {
+                        "variables": [
+                            {
+                                "name": "app",
+                                "value": "app-1"
+                            }
+                        ]
+                    }
+                }
+            },
+            "destination": {
+                "naming": {
+                    "suffix": "kfg-ex1"
+                },
+                "namespace": "default"
+            },
+            "reconciliation": {
+                "retryInterval": "10s",
+                "interval": "1m"
+            },
+            "sources": {
+                "flux": {
+                    "gitRepository": {
+                        "namespace": "default",
+                        "name": "example-configs"
+                    }
+                }
+            }
+        }
+    }
+
+    example_konfiguration = KonfigurationCR(fresh_kube_client, konfiguration_cr)
+
+    if example_konfiguration.exists():
+        example_konfiguration.update()
+    else:
+        example_konfiguration.create()
+
+    wait_for_objects_condition(
+        fresh_kube_client,
+        KonfigurationCR,
+        ["example-1"],
+        "default",
+        cr_ready,
+        120,
+        False,
+    )
+
+    cm_app_1_ex_1 = pykube.ConfigMap.objects(fresh_kube_client).get(name="app-1-kfg-ex1", namespace="default")
+
+    values_app_1_ex_1 = yaml.load(cm_app_1_ex_1.obj.get("data", {}).get("configmap-values.yaml", ""), SafeLoader)
+
+    assert values_app_1_ex_1.get("foo", "") == "override"
+    assert values_app_1_ex_1.get("bar", "") == "world"
+    assert values_app_1_ex_1.get("new", "") == "value"
+
+    assert values_app_1_ex_1.get("todo", {}).get("items", []) == ["item-1", "item-2", "item-3"]
+
+    secret_app_1_ex_1 = pykube.Secret.objects(fresh_kube_client).get(name="app-1-kfg-ex1", namespace="default")
+
+    secret_values_app_1_ex_1 = yaml.load(base64.b64decode(secret_app_1_ex_1.obj.get("data", {}).get("secret-values.yaml", "")), SafeLoader)
+
+    assert secret_values_app_1_ex_1.get("example") == "example"
+
+
+def install_crd(kube_cluster, relative_path_to_crd_file: str) -> None:
+    dir_path = os.path.dirname(os.path.realpath(__file__))
+
+    text = Path(f"{dir_path}/{relative_path_to_crd_file}").read_text()
+    dictionary = yaml.load(text, SafeLoader)
+
+    crd = pykube.CustomResourceDefinition(kube_cluster.kube_client, dictionary)
+
+    if crd.exists():
+        crd.update()
+    else:
+        crd.create()
+
+
+def cr_ready(cr: NamespacedKonfigureOperatorCR) -> bool:
     conditions = cr.obj.get("status", {}).get("conditions", [])
 
     for condition in conditions: