+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support Generalized Configuration System via new `Konfiguration` and `KonfigurationSchema` CRDs
+  - Bump `konfigure` to `v1.2.0`
+
 ## [0.8.0] - 2025-07-22
 
 ### Removed